+++ tests/native/native_pod_components.cpp
@@ -1,5 +1,10 @@
 #include "native_pod_components.h"
 
+float NativeIdentity(const vec2& value)
+{
+    return value.x;
+}
+
 namespace native_pod
 {
     float LengthSquared(vec2 value)
@@ -16,6 +21,22 @@ namespace native_pod
     {
         value.x += dx;
         value.y += dy;
+    }
+
+    float SumPointer(const vec2* value)
+    {
+        return value->x + value->y;
+    }
+
+    float DotPointerSelf(const vec2* value, const vec2& other)
+    {
+        return value->x * other.x + value->y * other.y;
+    }
+
+    void ScalePointer(vec2* value, float amount)
+    {
+        value->x *= amount;
+        value->y *= amount;
     }
 
     vec2 Make(float x, float y)